# fix(tsc): re-introduce global types removal check