The issue with the guard clause prior was that the out directory for the .d.ts file is likely to be a different place from the global types holder file. This meant the clause was not accurately detecting if it should remove the global types.

The guard clause is needed to some degree as large projects with many emitted files would run this removal which is fairly slow. vue-tsc is unusable without this guard clause.

Instead just get the file name from the file being written and the global types holder, if the dts file starts with the global types holder then perform the global types removal logic.

This allows for the output and source paths to be different. The removal may run more than once if similar names are present but it won't run on every file to be emitted.