@johnsoncodehk 2024-04-10
Thanks for the PR! Can you share how much the build time improved by introducing this condition in your case?