@blake-newman 2024-04-10
Our application: 
- 3340 (ts/x) files
- 851 vue files

Full build no guard check `4981s` (1 hour 23 min 1 second) 
Full build guard check `82s` (1 min 22 seconds)