@sawmurai 2024-04-11
Another data point: Our build slowed down by about 50% with the latest version.