@johnsoncodehk 2024-04-12
I haven't been able to reproduce the performance difference caused by this change. I would merge it first, but we'd better have a test case for this in case it breaks again in the future.