@messenjer 2024-04-19
For my project: 

Before 2.0.12:

```
> vue-tsc --build --force

npm run type-check  216.44s user 9.80s system 116% cpu 3:13.77 total
```

After 2.0.13:

```
> vue-tsc --build --force

npm run type-check  35.12s user 1.35s system 177% cpu 20.588 total
```

80% faster